Interface
Write a program using interface inheritance. Make two interfaces one is Shape2D
which contains a method to calculate the area of the circle and the second interface
Shape3D which contains method to calculate the volume of the sphere. Both the classes,
Circle and Sphere extends one abstract class Shape which contains a method display( ) to
display the area and volume.